refactor: extract error tests to sibling file
